requirements.txt
click==8.0.3
Flask==2.0.2
Flask-MySQLdb==0.2.0
Flask-WTF==1.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
mysql-connector-python==8.0.21
protobuf==3.13.0
six==1.15.0
Unidecode==1.3.2
Werkzeug==2.0.2
Whoosh==2.7.4
WTForms==3.0.1
wtforms-html5==0.6.1
markdown2
Flask-RESTful==0.3.9
PyJWT==2.6.0